requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
pyjwt
cryptography

 
